refactor: move GUI specific functions from asammdf.blocks.utils to asammdf.gui.serde
