+++ test/asammdf/gui/widgets/plot/test_PlotWidget_PushButtons.py
@@ -4,7 +4,7 @@ from PySide6 import QtCore, QtTest
 from PySide6.QtWidgets import QPushButton as QBtn
 from PySide6.QtWidgets import QTreeWidgetItemIterator
 
-from asammdf.blocks.utils import COLORS
+from asammdf.gui.serde import COLORS
 from asammdf.gui.utils import BLUE
 from test.asammdf.gui.test_base import Pixmap
 from test.asammdf.gui.widgets.test_BasePlotWidget import TestPlotWidget
